Added Condition for NO DATA report
If the Bluetooth dongle returns “NO DATA” string to indicate that there
was no report at the time, rather than crash, just go to the next thing
to read. Fortunately, nextStatToUpdate is a bitshifted integer, so it
was easy to do that, though we may miss some data. Better hacks will be
needed over time.